Don't create newline tokens
Instead, we will hackily use the `textPosition` field of the next
token when for example parsing `return` to determine if an expression
is to be returned or not. This means the `textPosition` field is
an essential part of the parse algorithm, and is under contract of the
lexer to be accurate, at least to the line number.

In practical terms, the `textPosition` field is going to stay around
for a while - not only to make it into the parse tree, but to be used
in errors, JSON-output parsing, recommendations, replacements, and
even added to debug output from LLVM.

However, this kills any notion of the parser being LL(k) or
representable in a CFG - at the very least, it's using non-token-
type data for its parse. In my mind this is a step that is taken
to be better than a rewriter, or a lexer hack. protosnirk parsing
is TDOP - the parser itself can be as hacky as it wants.

This is another straw on the great hay bale of "not being compatible
with general-purpose parsing tools" that hinders accessibility and
users' ability to hack on Snirk (such as autocomplete, syntax
highlighting, formatting, etc). It is my utmost hope that providing
a system-independent compiler service daemon with the compiler,
along with a walkthough of the parser code, will be enough.

Perhaps this will be rolled back at a later time, even to enable
additional hacks and tomfoolery. However, as long as none of Snirk's
parses ever become 'most vexing', the compiler is able to explain
its vexation, and the compiler service daemon is available, I will
be satisfied with the syntax of the language.